Added HUD elements for Fuel Added fuel mechanic for jetpack Moved cpp files into folders Added skeleton files for nodes
