+++ Config/DefaultInput.ini
@@ -1,11 +1,5 @@
 
 [/Script/Engine.InputSettings]
--AxisConfig=(AxisKeyName="Gamepad_LeftX",AxisProperties=(DeadZone=0.25,Exponent=1.f,Sensitivity=1.f))
--AxisConfig=(AxisKeyName="Gamepad_LeftY",AxisProperties=(DeadZone=0.25,Exponent=1.f,Sensitivity=1.f))
--AxisConfig=(AxisKeyName="Gamepad_RightX",AxisProperties=(DeadZone=0.25,Exponent=1.f,Sensitivity=1.f))
--AxisConfig=(AxisKeyName="Gamepad_RightY",AxisProperties=(DeadZone=0.25,Exponent=1.f,Sensitivity=1.f))
--AxisConfig=(AxisKeyName="MouseX",AxisProperties=(DeadZone=0.f,Exponent=1.f,Sensitivity=0.07f))
--AxisConfig=(AxisKeyName="MouseY",AxisProperties=(DeadZone=0.f,Exponent=1.f,Sensitivity=0.07f))
 -AxisConfig=(AxisKeyName="Gamepad_RightY",AxisProperties=(DeadZone=0.250000,Sensitivity=1.000000,Exponent=1.000000,bInvert=False))
 -AxisConfig=(AxisKeyName="MouseX",AxisProperties=(DeadZone=0.000000,Sensitivity=0.070000,Exponent=1.000000,bInvert=False))
 -AxisConfig=(AxisKeyName="MouseY",AxisProperties=(DeadZone=0.000000,Sensitivity=0.070000,Exponent=1.000000,bInvert=False))
@@ -62,11 +56,10 @@ DefaultViewportMouseLockMode=LockOnCapture
 -ActionMappings=(ActionName="Shoot",Key=LeftMouseButton,bShift=False,bCtrl=False,bAlt=False,bCmd=False)
 -ActionMappings=(ActionName="Jetpack",Key=Q,bShift=False,bCtrl=False,bAlt=False,bCmd=False)
 +ActionMappings=(ActionName="Shoot",Key=LeftMouseButton,bShift=False,bCtrl=False,bAlt=False,bCmd=False)
-+ActionMappings=(ActionName="Jetpack",Key=Q,bShift=False,bCtrl=False,bAlt=False,bCmd=False)
 -AxisMappings=(AxisName="MoveLeftRight",Key=A,Scale=-1.000000)
 -AxisMappings=(AxisName="MoveForwardBackward",Key=W,Scale=1.000000)
 -AxisMappings=(AxisName="MoveLeftRight",Key=D,Scale=1.000000)
--AxisMappings=(AxisName="MoveForwardBackward",Key=S,Scale=-0.750000)
+-AxisMappings=(AxisName="MoveForwardBackward",Key=S,Scale=-1.000000)
 -AxisMappings=(AxisName="MouseLookHorizontal",Key=MouseX,Scale=1.000000)
 -AxisMappings=(AxisName="MouseLookVertical",Key=MouseY,Scale=-1.000000)
 -AxisMappings=(AxisName="Jump",Key=SpaceBar,Scale=1.000000)
