--- Source/ProjectCovenant/Private/BattleRoyale.cpp
@@ -1,8 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#include "ProjectCovenant.h"
-#include "BattleRoyale.h"
-
-
-
-
