+++ Source/ProjectCovenant/Private/HUDs/PlayerHUD.cpp
@@ -1,0 +1,8 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "ProjectCovenant.h"
+#include "HUDs/PlayerHUD.h"
+
+
+
+
