+++ Source/ProjectCovenant/Private/Modes/BattleRoyale.cpp
@@ -1,0 +1,8 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "ProjectCovenant.h"
+#include "Modes/BattleRoyale.h"
+
+
+
+
