+++ Source/ProjectCovenant/Private/Nodes/Gas.cpp
@@ -1,0 +1,28 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "ProjectCovenant.h"
+#include "Gas.h"
+
+
+// Sets default values
+AGas::AGas()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+}
+
+// Called when the game starts or when spawned
+void AGas::BeginPlay()
+{
+	Super::BeginPlay();
+	
+}
+
+// Called every frame
+void AGas::Tick(float DeltaTime)
+{
+	Super::Tick(DeltaTime);
+
+}
+
