+++ Source/ProjectCovenant/Private/Nodes/Metal.cpp
@@ -1,0 +1,28 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "ProjectCovenant.h"
+#include "Metal.h"
+
+
+// Sets default values
+AMetal::AMetal()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+}
+
+// Called when the game starts or when spawned
+void AMetal::BeginPlay()
+{
+	Super::BeginPlay();
+	
+}
+
+// Called every frame
+void AMetal::Tick(float DeltaTime)
+{
+	Super::Tick(DeltaTime);
+
+}
+
