+++ Source/ProjectCovenant/Private/Nodes/Organic.cpp
@@ -1,0 +1,28 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "ProjectCovenant.h"
+#include "Organic.h"
+
+
+// Sets default values
+AOrganic::AOrganic()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+}
+
+// Called when the game starts or when spawned
+void AOrganic::BeginPlay()
+{
+	Super::BeginPlay();
+	
+}
+
+// Called every frame
+void AOrganic::Tick(float DeltaTime)
+{
+	Super::Tick(DeltaTime);
+
+}
+
