--- Source/ProjectCovenant/Public/Deathmatch.h
@@ -1,19 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "GameFramework/GameMode.h"
-#include "Deathmatch.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class PROJECTCOVENANT_API ADeathmatch : public AGameMode
-{
-	GENERATED_BODY()
-	
-	
-	
-	
-};
