+++ Source/ProjectCovenant/Public/HUDs/PlayerHUD.h
@@ -1,0 +1,19 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "GameFramework/HUD.h"
+#include "PlayerHUD.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PROJECTCOVENANT_API APlayerHUD : public AHUD
+{
+	GENERATED_BODY()
+	
+	
+	
+	
+};
