+++ Source/ProjectCovenant/Public/Modes/BattleRoyale.h
@@ -1,0 +1,19 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "GameFramework/GameMode.h"
+#include "BattleRoyale.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PROJECTCOVENANT_API ABattleRoyale : public AGameMode
+{
+	GENERATED_BODY()
+	
+	
+	
+	
+};
