+++ Source/ProjectCovenant/Public/Modes/Covenant.h
@@ -1,0 +1,19 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "GameFramework/GameMode.h"
+#include "Covenant.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PROJECTCOVENANT_API ACovenant : public AGameMode
+{
+	GENERATED_BODY()
+	
+	
+	
+	
+};
